Start moving shared cancellation survey functionality to the same location as the survey component.
