+++ client/me/purchases/cancel-purchase/button.jsx
@@ -18,8 +18,8 @@ import { clearPurchases } from 'state/purchases/actions';
 import { connect } from 'react-redux';
 import Dialog from 'components/dialog';
 import CancelPurchaseForm from 'components/marketing-survey/cancel-purchase-form';
+import enrichedSurveyData from 'components/marketing-survey/cancel-purchase-form/enrichedSurveyData';
 import { getName, getSubscriptionEndDate, isOneTimePurchase, isRefundable, isSubscription } from 'lib/purchases';
-import { enrichedSurveyData } from '../utils';
 import { isDomainRegistration, isJetpackPlan, isBusiness } from 'lib/products-values';
 import notices from 'notices';
 import paths from 'me/purchases/paths';
